requirements.txt
# Tunisia Smart Tourism — Local Requirements
# Install with: pip install -r requirements.txt

# Database
sqlalchemy==2.0.30

# ML Stack
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
torch==2.3.0
joblib==1.4.2

# API calls
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
